fix(phase10): S8 Gemini - deepcopy export modules (concurrency safety)
Gemini code final-pass raised 1 CRITICAL (verified valid; a genuine
cross-family find): _OnnxForward toggled onnx_export on the SHARED
backbone from _predict_module(), so a concurrent predict() during an
export_onnx trace would skip the eager validation guards and take the
ONNX LSTM path - the Phase 9 G-C1 statefulness-under-concurrent-
predict hazard class, missed by the same-family S7 swarm.

Fix: _OnnxForward.__init__ now copy.deepcopy's backbone+head and sets
onnx_export=True on the PRIVATE clone permanently; the shared
estimator modules are never mutated and forward() no longer toggles
shared state (no try/finally). Export is one-shot so the deepcopy
cost is irrelevant. Parity + op-surface tests stay green (clone has
identical weights). Also applied Gemini's NITPICK: dropped the
redundant `.to(torch.int64)` (bool sum is already int64), removing a
needless ONNX Cast. Gemini's IMPROVEMENT independently CONFIRMED the
modular-roll == stable-argsort equivalence (no change needed).

Per the gemini-final-pass protocol a new valid CRITICAL invalidates
S7 consensus; one more Claude /review round on this fix follows.

+++ src/seq_sklearn/models/transformer/tft/backbone.py
@@ -174,7 +174,7 @@ class TFTBackbone(BaseBackbone):
             # the leading pad block (in order). Identical to the
             # stable-argsort result on left-padded input; only Range/
             # Add/Sub/Mod ops, all ONNX-safe.
-            roll = (seq_len - lengths).to(torch.int64)  # (B,)
+            roll = seq_len - lengths  # (B,); lengths is int64 (bool sum)
             pos = torch.arange(seq_len, device=mask.device)  # (L,)
             order = (pos[None, :] + roll[:, None]) % seq_len  # (B, L)
             inverse = (pos[None, :] - roll[:, None]) % seq_len  # (B, L)
